+++ test/e7_connection_conditions.js
@@ -1,7 +1,7 @@
 describe(require('path').basename(__filename), function () {
 
-  require('benchmarket').start();
-  after(require('benchmarket').store());
+  //require('benchmarket').start();
+  //after(//require('benchmarket').store());
 
   var expect = require('expect.js');
   var happn = require('../lib/index');
@@ -104,6 +104,6 @@ describe(require('path').basename(__filename), function () {
 
   });
 
-  require('benchmarket').stop();
+  //require('benchmarket').stop();
 
 });