+++ test/e8_manual_destroy.js
@@ -1,7 +1,7 @@
 describe('e8_manual_destroy', function () {
 
-  require('benchmarket').start();
-  after(require('benchmarket').store());
+  //require('benchmarket').start();
+  //after(//require('benchmarket').store());
 
   var expect = require('expect.js');
   var happn = require('../lib/index');
@@ -69,6 +69,6 @@ describe('e8_manual_destroy', function () {
 
   });
 
-  require('benchmarket').stop();
+  //require('benchmarket').stop();
 
 });